fixed data population logic

+++ app/migrations/000006_create_review_tags.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS ReviewTags (
   id INT PRIMARY KEY,
-  jan INT,
+  jan BIGINT,
   tag_from_review VARCHAR(255),
   FOREIGN KEY (jan) REFERENCES Products(jan)
 );
